pipeline/requirements.txt
feedparser==6.0.11
httpx==0.27.0
groq==0.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
